requirements.txt
starlette
uvicorn
mcp
openai
transformers
peft
Pillow
numpy
pyyaml
